+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Private constructor in struct. [#79](https://github.com/nojaf/telplin/issues/79)
 * Generic type parameter doesn't parse internally. [#68](https://github.com/nojaf/telplin/issues/68)
 * Optional function type cannot be parsed. [#78](https://github.com/nojaf/telplin/issues/78)
+* Property member with function return type. [#66](https://github.com/nojaf/telplin/issues/66)
 
 ## 0.7.0 - 05-06-23